demo/backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
google-genai==1.5.0
pymilvus[milvus-lite]
pydantic==2.10.4
python-dotenv==1.0.1
